重构WfClientProxy的机制，引入基类，更为好用。 增加UserOperationLog的WCF方法

+++ Framework/MCS.Library.SOA.DataObjects/Log/UserOperationLogAdapter.cs
@@ -11,64 +11,64 @@ using MCS.Library.SOA.DataObjects.Workflow;
 
 namespace MCS.Library.SOA.DataObjects
 {
-	public class UserOperationLogAdapter : UpdatableAndLoadableAdapterBase<UserOperationLog, UserOperationLogCollection>
-	{
-		public static readonly UserOperationLogAdapter Instance = new UserOperationLogAdapter();
+    public class UserOperationLogAdapter : UpdatableAndLoadableAdapterBase<UserOperationLog, UserOperationLogCollection>
+    {
+        public static readonly UserOperationLogAdapter Instance = new UserOperationLogAdapter();
 
-		private UserOperationLogAdapter()
-		{
-		}
+        private UserOperationLogAdapter()
+        {
+        }
 
-		/// <summary>
-		/// 添加一行数据，并返回新主键
-		/// </summary>
-		/// <param name="data"></param>
-		/// <returns></returns>
-		public int InsertData(UserOperationLog data)
-		{
-			ORMappingItemCollection mappings = ORMapping.GetMappingInfo<UserOperationLog>();
+        /// <summary>
+        /// 添加一行数据，并返回新主键
+        /// </summary>
+        /// <param name="data"></param>
+        /// <returns></returns>
+        public int InsertData(UserOperationLog data)
+        {
+            ORMappingItemCollection mappings = ORMapping.GetMappingInfo<UserOperationLog>();
 
-			string sql = ORMapping.GetInsertSql(data, mappings, TSqlBuilder.Instance);
+            string sql = ORMapping.GetInsertSql(data, mappings, TSqlBuilder.Instance);
 
-			decimal result = (decimal)DbHelper.RunSqlReturnScalar(string.Format("{0} \n SELECT @@IDENTITY", sql), this.GetConnectionName());
+            decimal result = (decimal)DbHelper.RunSqlReturnScalar(string.Format("{0} \n SELECT @@IDENTITY", sql), this.GetConnectionName());
 
-			return decimal.ToInt32(result);
-		}
+            return decimal.ToInt32(result);
+        }
 
-		public UserOperationLog Load(int id)
-		{
-			UserOperationLogCollection logCollection = Load(builder => builder.AppendItem("ID", id));
+        public UserOperationLog Load(Int64 id)
+        {
+            UserOperationLogCollection logCollection = Load(builder => builder.AppendItem("ID", id));
 
-			(logCollection.Count > 0).FalseThrow("不能找到ID为{0}的USER_OPERATION_LOG记录", id);
+            (logCollection.Count > 0).FalseThrow("不能找到ID为{0}的USER_OPERATION_LOG记录", id);
 
-			return logCollection[0];
-		}
+            return logCollection[0];
+        }
 
-		public UserOperationLogCollection LoadByResourceID(string resourceID)
-		{
-			resourceID.CheckStringIsNullOrEmpty("resourceID");
+        public UserOperationLogCollection LoadByResourceID(string resourceID)
+        {
+            resourceID.CheckStringIsNullOrEmpty("resourceID");
 
-			UserOperationLogCollection result = Load(builder => builder.AppendItem("RESOURCE_ID", resourceID));
+            UserOperationLogCollection result = Load(builder => builder.AppendItem("RESOURCE_ID", resourceID));
 
-			result.Sort((l1, l2) =>
-			{
-				int r = 0;
+            result.Sort((l1, l2) =>
+            {
+                int r = 0;
 
-				if (l1.OperationDateTime > l2.OperationDateTime)
-					r = -1;
-				else
-					if (l1.OperationDateTime < l2.OperationDateTime)
-						r = 1;
+                if (l1.OperationDateTime > l2.OperationDateTime)
+                    r = -1;
+                else
+                    if (l1.OperationDateTime < l2.OperationDateTime)
+                        r = 1;
 
-				return r;
-			});
+                return r;
+            });
 
-			return result;
-		}
+            return result;
+        }
 
-		protected override string GetConnectionName()
-		{
-			return WfRuntime.ProcessContext.SimulationContext.GetConnectionName(AppLogSettings.GetConfig().ConnectionName);
-		}
-	}
+        protected override string GetConnectionName()
+        {
+            return WfRuntime.ProcessContext.SimulationContext.GetConnectionName(AppLogSettings.GetConfig().ConnectionName);
+        }
+    }
 }
